change long long to int64_t type and unsigned long long to uint64_t type

+++ Cryptography/diffie-hellman.c
@@ -9,22 +9,22 @@
 int dh_private_key_a;
 int dh_private_key_b;
 
-long long DH_generate_keyPair(long long p, int alpha) {
+int64_t DH_generate_keyPair(int64_t p, int alpha) {
 	srand((unsigned int) time(NULL));
 	dh_private_key_a = rand();
 	dh_private_key_b = rand();
 	return exponent(alpha, dh_private_key_a, p);
 }
 
-void DH_encrypt(long long p, int alpha, long long Kpub, int message, long long *ephemeral_key, long long *cipher) {
+void DH_encrypt(int64_t p, int alpha, int64_t Kpub, int message, int64_t *ephemeral_key, int64_t *cipher) {
 	*ephemeral_key = exponent(alpha, dh_private_key_b, p);
-	long long Km = exponent(Kpub, dh_private_key_b, p);
+	int64_t Km = exponent(Kpub, dh_private_key_b, p);
 	*cipher = (message * Km) % p;
 }
 
-void DH_decrypt(long long p, int alpha, long long Keph, long long cipher, long long *message) {
-	long long km = exponent(Keph, dh_private_key_a, p);
-	long long *km_inverse = inverse(km, p);
+void DH_decrypt(int64_t p, int alpha, int64_t Keph, int64_t cipher, int64_t *message) {
+	int64_t km = exponent(Keph, dh_private_key_a, p);
+	int64_t *km_inverse = inverse(km, p);
 	if (km_inverse == NULL) return;
 	*message = (cipher * *km_inverse) % p;
 	free(km_inverse);
