+++ Cryptography/diffie-hellman.h
@@ -9,16 +9,17 @@
 #define diffie_hellman_h
 
 #include <stdio.h>
+#include <stdint.h>
 #include <time.h>
 #include "inverse.h"
 #include "square-and-multiply.h"
 
 #define BIG_PRIME 2291447117
 
-long long DH_generate_keyPair(long long p, int alpha);
+int64_t DH_generate_keyPair(int64_t p, int alpha);
 
-void DH_encrypt(long long p, int alpha, long long Kpub, int message, long long *ephemeral_key, long long *cipher);
+void DH_encrypt(int64_t p, int alpha, int64_t Kpub, int message, int64_t *ephemeral_key, int64_t *cipher);
 
-void DH_decrypt(long long p, int alpha, long long Keph, long long cipher, long long *message);
+void DH_decrypt(int64_t p, int alpha, int64_t Keph, int64_t cipher, int64_t *message);
 
 #endif /* diffie_hellman_h */
