+++ Cryptography/inverse.c
@@ -18,7 +18,7 @@ void scan_for_inverse(int *a, int *n) {
 void print_inverse(int a, int n) {
 	int g = gcd(a, n);
 	printf("%dx === %d (mod %d)\n", a, g, n);
-	long long *inv = inverse(a, n);
+	int64_t *inv = inverse(a, n);
 	if (inv == NULL) {
 		printf("Can't find the inverse between %d and %d\n", a, n);
 		printf("because the gcd between them is %d\n", g);
@@ -29,12 +29,12 @@ void print_inverse(int a, int n) {
 	free(inv);
 }
 
-long long *inverse(long long a, long long n) {
-	long long gcd, x, y;
+int64_t *inverse(int64_t a, int64_t n) {
+	int64_t gcd, x, y;
 	extended_gcd(a, n, &gcd, &x, &y);
 	if (gcd != 1) return NULL;
 	a %= n;
-	long long *toRet = malloc(sizeof(long long));
+	int64_t *toRet = malloc(sizeof(int64_t));
 	*toRet = (x < 0) ? x + n : x;
 	return toRet;
 }
