+++ Cryptography/inverse.h
@@ -10,6 +10,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include "gcd.h"
 
 void scan_for_inverse(int *a, int *n);
@@ -18,6 +19,6 @@ void print_inverse(int a, int n);
 
 void print_inverse_error(int a, int n, int gcd);
 
-long long *inverse(long long a, long long n);
+int64_t *inverse(int64_t a, int64_t n);
 
 #endif /* inverse_h */
